feat: resolve comments to CLI experience (#10)
