+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/_show.py
@@ -18,7 +18,7 @@ class Show(AAZCommand):
     """Get details of the API.
 
     :example: Show API details
-        az apic api show -g contoso-resources -s contoso --name echo-api
+        az apic api show -g contoso-resources -s contoso --api-id echo-api
     """
 
     _aaz_info = {
@@ -44,9 +44,9 @@ class Show(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.api_name = AAZStrArg(
-            options=["--api", "--name", "--api-name"],
-            help="The name of the API.",
+        _args_schema.api_id = AAZStrArg(
+            options=["--api-id"],
+            help="The id of the API.",
             required=True,
             id_part="child_name_2",
             fmt=AAZStrArgFormat(
@@ -127,7 +127,7 @@ class Show(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "apiName", self.ctx.args.api_name,
+                    "apiName", self.ctx.args.api_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -218,6 +218,7 @@ class Show(AAZCommand):
             properties.license = AAZObjectType()
             properties.lifecycle_stage = AAZStrType(
                 serialized_name="lifecycleStage",
+                flags={"read_only": True},
             )
             properties.summary = AAZStrType()
             properties.terms_of_service = AAZObjectType(
