+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/_update.py
@@ -15,10 +15,10 @@ from azure.cli.core.aaz import *
     "apic api update",
 )
 class Update(AAZCommand):
-    """Update new or updates existing API.
+    """Update existing API.
 
     :example: Update API
-        az apic api update -g contoso-resources -s contoso --name echo-api --summary "Basic REST API service"
+        az apic api update -g contoso-resources -s contoso --api-id echo-api --summary "Basic REST API service"
     """
 
     _aaz_info = {
@@ -46,9 +46,9 @@ class Update(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.api_name = AAZStrArg(
-            options=["--api", "--name", "--api-name"],
-            help="The name of the API.",
+        _args_schema.api_id = AAZStrArg(
+            options=["--api-id"],
+            help="The id of the API.",
             required=True,
             id_part="child_name_2",
             fmt=AAZStrArgFormat(
@@ -111,10 +111,10 @@ class Update(AAZCommand):
             help="Additional, external documentation for the API.",
             nullable=True,
         )
-        _args_schema.kind = AAZStrArg(
-            options=["--kind"],
+        _args_schema.type = AAZStrArg(
+            options=["--type"],
             arg_group="Properties",
-            help="Kind of API. For example, REST or GraphQL.",
+            help="Type of API.",
             enum={"graphql": "graphql", "grpc": "grpc", "rest": "rest", "soap": "soap", "webhook": "webhook", "websocket": "websocket"},
         )
         _args_schema.license = AAZObjectArg(
@@ -305,7 +305,7 @@ class Update(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "apiName", self.ctx.args.api_name,
+                    "apiName", self.ctx.args.api_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -396,7 +396,7 @@ class Update(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "apiName", self.ctx.args.api_name,
+                    "apiName", self.ctx.args.api_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -488,7 +488,7 @@ class Update(AAZCommand):
                 properties.set_prop("customProperties", AAZFreeFormDictType, ".custom_properties")
                 properties.set_prop("description", AAZStrType, ".description")
                 properties.set_prop("externalDocumentation", AAZListType, ".external_documentation")
-                properties.set_prop("kind", AAZStrType, ".kind", typ_kwargs={"flags": {"required": True}})
+                properties.set_prop("kind", AAZStrType, ".type", typ_kwargs={"flags": {"required": True}})
                 properties.set_prop("license", AAZObjectType, ".license")
                 properties.set_prop("summary", AAZStrType, ".summary")
                 properties.set_prop("termsOfService", AAZObjectType, ".terms_of_service")
@@ -589,6 +589,7 @@ class _UpdateHelper:
         properties.license = AAZObjectType()
         properties.lifecycle_stage = AAZStrType(
             serialized_name="lifecycleStage",
+            flags={"read_only": True},
         )
         properties.summary = AAZStrType()
         properties.terms_of_service = AAZObjectType(
