+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/definition/__cmd_group.py
@@ -15,7 +15,7 @@ from azure.cli.core.aaz import *
     "apic api definition",
 )
 class __CMDGroup(AAZCommandGroup):
-    """Commands to manage API Definitions in API Center
+    """Commands to manage API definitions in API Center
     """
     pass
