+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/definition/__init__.py
@@ -9,10 +9,10 @@
 # flake8: noqa
 
 from .__cmd_group import *
+from ._check_exists import *
 from ._create import *
 from ._delete import *
 from ._export_specification import *
-from ._head import *
 from ._import_specification import *
 from ._list import *
 from ._show import *
