+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/definition/_check_exists.py
@@ -12,16 +12,16 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "apic metadata-schema head",
+    "apic api definition check-exists",
 )
-class Head(AAZCommand):
-    """Checks if specified metadata schema exists.
+class CheckExists(AAZCommand):
+    """Checks if specified API definition exists.
     """
 
     _aaz_info = {
         "version": "2024-03-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.apicenter/services/{}/metadataschemas/{}", "2024-03-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.apicenter/services/{}/workspaces/{}/apis/{}/versions/{}/definitions/{}", "2024-03-01"],
         ]
     }
 
@@ -41,11 +41,21 @@ class Head(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.metadata_schema_name = AAZStrArg(
-            options=["--name", "--metadata-schema", "--metadata-schema-name"],
-            help="The name of the metadata schema.",
+        _args_schema.api_id = AAZStrArg(
+            options=["--api-id"],
+            help="The id of the API.",
             required=True,
-            id_part="child_name_1",
+            id_part="child_name_2",
+            fmt=AAZStrArgFormat(
+                max_length=90,
+                min_length=1,
+            ),
+        )
+        _args_schema.definition_id = AAZStrArg(
+            options=["--definition-id"],
+            help="The id of the API definition.",
+            required=True,
+            id_part="child_name_4",
             fmt=AAZStrArgFormat(
                 max_length=90,
                 min_length=1,
@@ -64,11 +74,32 @@ class Head(AAZCommand):
                 min_length=1,
             ),
         )
+        _args_schema.version_id = AAZStrArg(
+            options=["--version-id"],
+            help="The id of the API version.",
+            required=True,
+            id_part="child_name_3",
+            fmt=AAZStrArgFormat(
+                max_length=90,
+                min_length=1,
+            ),
+        )
+        _args_schema.workspace_name = AAZStrArg(
+            options=["-w", "--workspace", "--workspace-name"],
+            help="The name of the workspace.",
+            required=True,
+            id_part="child_name_1",
+            default="default",
+            fmt=AAZStrArgFormat(
+                max_length=90,
+                min_length=1,
+            ),
+        )
         return cls._args_schema
 
     def _execute_operations(self):
         self.pre_operations()
-        self.MetadataSchemasHead(ctx=self.ctx)()
+        self.ApiDefinitionsHead(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -79,7 +110,7 @@ class Head(AAZCommand):
     def post_operations(self):
         pass
 
-    class MetadataSchemasHead(AAZHttpOperation):
+    class ApiDefinitionsHead(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
@@ -93,7 +124,7 @@ class Head(AAZCommand):
         @property
         def url(self):
             return self.client.format_url(
-                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ApiCenter/services/{serviceName}/metadataSchemas/{metadataSchemaName}",
+                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ApiCenter/services/{serviceName}/workspaces/{workspaceName}/apis/{apiName}/versions/{versionName}/definitions/{definitionName}",
                 **self.url_parameters
             )
 
@@ -109,7 +140,11 @@ class Head(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "metadataSchemaName", self.ctx.args.metadata_schema_name,
+                    "apiName", self.ctx.args.api_id,
+                    required=True,
+                ),
+                **self.serialize_url_param(
+                    "definitionName", self.ctx.args.definition_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -122,6 +157,14 @@ class Head(AAZCommand):
                 ),
                 **self.serialize_url_param(
                     "subscriptionId", self.ctx.subscription_id,
+                    required=True,
+                ),
+                **self.serialize_url_param(
+                    "versionName", self.ctx.args.version_id,
+                    required=True,
+                ),
+                **self.serialize_url_param(
+                    "workspaceName", self.ctx.args.workspace_name,
                     required=True,
                 ),
             }
@@ -141,8 +184,8 @@ class Head(AAZCommand):
             pass
 
 
-class _HeadHelper:
-    """Helper class for Head"""
+class _CheckExistsHelper:
+    """Helper class for CheckExists"""
 
 
-__all__ = ["Head"]
+__all__ = ["CheckExists"]
