+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/definition/_list.py
@@ -18,7 +18,7 @@ class List(AAZCommand):
     """List a collection of API definitions.
 
     :example: List API definitions
-        az apic api definition list -g api-center-test -s contosoeuap --api-name echo-api --version 2023-01-01
+        az apic api definition list -g api-center-test -s contosoeuap --api-id echo-api --version-id 2023-01-01
     """
 
     _aaz_info = {
@@ -45,9 +45,9 @@ class List(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.api_name = AAZStrArg(
-            options=["--api", "--api-name"],
-            help="The name of the API.",
+        _args_schema.api_id = AAZStrArg(
+            options=["--api-id"],
+            help="The id of the API.",
             required=True,
             fmt=AAZStrArgFormat(
                 max_length=90,
@@ -66,9 +66,9 @@ class List(AAZCommand):
                 min_length=1,
             ),
         )
-        _args_schema.version_name = AAZStrArg(
-            options=["--version", "--version-name"],
-            help="The name of the API version.",
+        _args_schema.version_id = AAZStrArg(
+            options=["--version-id"],
+            help="The id of the API version.",
             required=True,
             fmt=AAZStrArgFormat(
                 max_length=90,
@@ -139,7 +139,7 @@ class List(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "apiName", self.ctx.args.api_name,
+                    "apiName", self.ctx.args.api_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -155,7 +155,7 @@ class List(AAZCommand):
                     required=True,
                 ),
                 **self.serialize_url_param(
-                    "versionName", self.ctx.args.version_name,
+                    "versionName", self.ctx.args.version_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
