+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/definition/_update.py
@@ -15,10 +15,10 @@ from azure.cli.core.aaz import *
     "apic api definition update",
 )
 class Update(AAZCommand):
-    """Update new or updates existing API definition.
+    """Update existing API definition.
 
     :example: Update API definition
-        az apic api definition update -g api-center-test -s contosoeuap --api-name echo-api --version 2023-01-01 --name "openapi" --title "OpenAPI"
+        az apic api definition update -g api-center-test -s contosoeuap --api-id echo-api --version-id 2023-01-01 --definition-id "openapi" --title "OpenAPI"
     """
 
     _aaz_info = {
@@ -46,9 +46,9 @@ class Update(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.api_name = AAZStrArg(
-            options=["--api", "--api-name"],
-            help="The name of the API.",
+        _args_schema.api_id = AAZStrArg(
+            options=["--api-id"],
+            help="The id of the API.",
             required=True,
             id_part="child_name_2",
             fmt=AAZStrArgFormat(
@@ -56,9 +56,9 @@ class Update(AAZCommand):
                 min_length=1,
             ),
         )
-        _args_schema.definition_name = AAZStrArg(
-            options=["--name", "--definition", "--definition-name"],
-            help="The name of the API definition.",
+        _args_schema.definition_id = AAZStrArg(
+            options=["--definition-id"],
+            help="The id of the API definition.",
             required=True,
             id_part="child_name_4",
             fmt=AAZStrArgFormat(
@@ -79,9 +79,9 @@ class Update(AAZCommand):
                 min_length=1,
             ),
         )
-        _args_schema.version_name = AAZStrArg(
-            options=["--version", "--version-name"],
-            help="The name of the API version.",
+        _args_schema.version_id = AAZStrArg(
+            options=["--version-id"],
+            help="The id of the API version.",
             required=True,
             id_part="child_name_3",
             fmt=AAZStrArgFormat(
@@ -180,11 +180,11 @@ class Update(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "apiName", self.ctx.args.api_name,
+                    "apiName", self.ctx.args.api_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
-                    "definitionName", self.ctx.args.definition_name,
+                    "definitionName", self.ctx.args.definition_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -200,7 +200,7 @@ class Update(AAZCommand):
                     required=True,
                 ),
                 **self.serialize_url_param(
-                    "versionName", self.ctx.args.version_name,
+                    "versionName", self.ctx.args.version_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -279,11 +279,11 @@ class Update(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "apiName", self.ctx.args.api_name,
+                    "apiName", self.ctx.args.api_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
-                    "definitionName", self.ctx.args.definition_name,
+                    "definitionName", self.ctx.args.definition_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -299,7 +299,7 @@ class Update(AAZCommand):
                     required=True,
                 ),
                 **self.serialize_url_param(
-                    "versionName", self.ctx.args.version_name,
+                    "versionName", self.ctx.args.version_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
