+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/deployment/_show.py
@@ -18,7 +18,7 @@ class Show(AAZCommand):
     """Get details of the API deployment.
 
     :example: Show API deployment details
-        az apic api deployment show -g api-center-test -s contoso --name production --api-name echo-api
+        az apic api deployment show -g api-center-test -s contoso --deployment-id production --api-id echo-api
     """
 
     _aaz_info = {
@@ -44,9 +44,9 @@ class Show(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.api_name = AAZStrArg(
-            options=["--api", "--api-name"],
-            help="The name of the API.",
+        _args_schema.api_id = AAZStrArg(
+            options=["--api-id"],
+            help="The id of the API.",
             required=True,
             id_part="child_name_2",
             fmt=AAZStrArgFormat(
@@ -54,9 +54,9 @@ class Show(AAZCommand):
                 min_length=1,
             ),
         )
-        _args_schema.deployment_name = AAZStrArg(
-            options=["--name", "--deployment", "--deployment-name"],
-            help="The name of the API deployment.",
+        _args_schema.deployment_id = AAZStrArg(
+            options=["--deployment-id"],
+            help="The id of the API deployment.",
             required=True,
             id_part="child_name_3",
             fmt=AAZStrArgFormat(
@@ -137,11 +137,11 @@ class Show(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "apiName", self.ctx.args.api_name,
+                    "apiName", self.ctx.args.api_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
-                    "deploymentName", self.ctx.args.deployment_name,
+                    "deploymentName", self.ctx.args.deployment_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
