+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/api/version/__cmd_group.py
@@ -15,7 +15,7 @@ from azure.cli.core.aaz import *
     "apic api version",
 )
 class __CMDGroup(AAZCommandGroup):
-    """Commands to manage API Versions in API Center
+    """Commands to manage API versions in API Center
     """
     pass
