+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/environment/__cmd_group.py
@@ -15,7 +15,7 @@ from azure.cli.core.aaz import *
     "apic environment",
 )
 class __CMDGroup(AAZCommandGroup):
-    """Commands to manage Environments in API Center
+    """Commands to manage environments in API Center
     """
     pass
