+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/environment/__init__.py
@@ -9,9 +9,9 @@
 # flake8: noqa
 
 from .__cmd_group import *
+from ._check_exists import *
 from ._create import *
 from ._delete import *
-from ._head import *
 from ._list import *
 from ._show import *
 from ._update import *
