+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/environment/_check_exists.py
@@ -12,23 +12,16 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "apic metadata-schema delete",
-    confirmation="Are you sure you want to perform this operation?",
+    "apic environment check-exists",
 )
-class Delete(AAZCommand):
-    """Delete specified metadata schema.
-
-    :example: Delete Metadata Schema
-        az az apic metadata-schema delete --resource-group api-center-test --service-name contoso --name "test1"
-
-    :example: Delete schema
-        az apic metadata-schema delete -g api-center-test -s contosoeuap --name "approver"
+class CheckExists(AAZCommand):
+    """Checks if specified environment exists.
     """
 
     _aaz_info = {
         "version": "2024-03-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.apicenter/services/{}/metadataschemas/{}", "2024-03-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.apicenter/services/{}/workspaces/{}/environments/{}", "2024-03-01"],
         ]
     }
 
@@ -48,11 +41,11 @@ class Delete(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.metadata_schema_name = AAZStrArg(
-            options=["--name", "--metadata-schema", "--metadata-schema-name"],
-            help="The name of the metadata schema.",
+        _args_schema.environment_id = AAZStrArg(
+            options=["--environment-id"],
+            help="The id of the environment.",
             required=True,
-            id_part="child_name_1",
+            id_part="child_name_2",
             fmt=AAZStrArgFormat(
                 max_length=90,
                 min_length=1,
@@ -71,11 +64,22 @@ class Delete(AAZCommand):
                 min_length=1,
             ),
         )
+        _args_schema.workspace_name = AAZStrArg(
+            options=["-w", "--workspace", "--workspace-name"],
+            help="The name of the workspace.",
+            required=True,
+            id_part="child_name_1",
+            default="default",
+            fmt=AAZStrArgFormat(
+                max_length=90,
+                min_length=1,
+            ),
+        )
         return cls._args_schema
 
     def _execute_operations(self):
         self.pre_operations()
-        self.MetadataSchemasDelete(ctx=self.ctx)()
+        self.EnvironmentsHead(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -86,7 +90,7 @@ class Delete(AAZCommand):
     def post_operations(self):
         pass
 
-    class MetadataSchemasDelete(AAZHttpOperation):
+    class EnvironmentsHead(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
@@ -94,21 +98,19 @@ class Delete(AAZCommand):
             session = self.client.send_request(request=request, stream=False, **kwargs)
             if session.http_response.status_code in [200]:
                 return self.on_200(session)
-            if session.http_response.status_code in [204]:
-                return self.on_204(session)
 
             return self.on_error(session.http_response)
 
         @property
         def url(self):
             return self.client.format_url(
-                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ApiCenter/services/{serviceName}/metadataSchemas/{metadataSchemaName}",
+                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ApiCenter/services/{serviceName}/workspaces/{workspaceName}/environments/{environmentName}",
                 **self.url_parameters
             )
 
         @property
         def method(self):
-            return "DELETE"
+            return "HEAD"
 
         @property
         def error_format(self):
@@ -118,7 +120,7 @@ class Delete(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "metadataSchemaName", self.ctx.args.metadata_schema_name,
+                    "environmentName", self.ctx.args.environment_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -131,6 +133,10 @@ class Delete(AAZCommand):
                 ),
                 **self.serialize_url_param(
                     "subscriptionId", self.ctx.subscription_id,
+                    required=True,
+                ),
+                **self.serialize_url_param(
+                    "workspaceName", self.ctx.args.workspace_name,
                     required=True,
                 ),
             }
@@ -149,12 +155,9 @@ class Delete(AAZCommand):
         def on_200(self, session):
             pass
 
-        def on_204(self, session):
-            pass
+
+class _CheckExistsHelper:
+    """Helper class for CheckExists"""
 
 
-class _DeleteHelper:
-    """Helper class for Delete"""
-
-
-__all__ = ["Delete"]
+__all__ = ["CheckExists"]
