+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/environment/_create.py
@@ -15,10 +15,10 @@ from azure.cli.core.aaz import *
     "apic environment create",
 )
 class Create(AAZCommand):
-    """Create new or updates existing environment.
+    """Create a new environment.
 
     :example: Create environment
-        az apic environment create -g api-center-test -s contosoeuap --name public --title "Public cloud" --kind "development"
+        az apic environment create -g api-center-test -s contosoeuap --environment-id public --title "Public cloud" --type "development"
     """
 
     _aaz_info = {
@@ -44,9 +44,9 @@ class Create(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.environment_name = AAZStrArg(
-            options=["--name", "--environment", "--environment-name"],
-            help="The name of the environment.",
+        _args_schema.environment_id = AAZStrArg(
+            options=["--environment-id"],
+            help="The id of the environment.",
             required=True,
             fmt=AAZStrArgFormat(
                 max_length=90,
@@ -88,18 +88,19 @@ class Create(AAZCommand):
         _args_schema.description = AAZStrArg(
             options=["--description"],
             arg_group="Properties",
-            help="Description.",
+            help="Environment description.",
         )
-        _args_schema.kind = AAZStrArg(
-            options=["--kind"],
+        _args_schema.type = AAZStrArg(
+            options=["--type"],
             arg_group="Properties",
-            help="Environment kind.",
             required=True,
+            help="Environment type.",
             enum={"development": "development", "production": "production", "staging": "staging", "testing": "testing"},
         )
         _args_schema.onboarding = AAZObjectArg(
             options=["--onboarding"],
             arg_group="Properties",
+            help="Provide onboarding documentation related to your environment, e.g. {developerPortalUri:['https://developer.contoso.com'],instructions:'instructions markdown'}",
         )
         _args_schema.server = AAZObjectArg(
             options=["--server"],
@@ -190,7 +191,7 @@ class Create(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "environmentName", self.ctx.args.environment_name,
+                    "environmentName", self.ctx.args.environment_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -247,7 +248,7 @@ class Create(AAZCommand):
             if properties is not None:
                 properties.set_prop("customProperties", AAZFreeFormDictType, ".custom_properties")
                 properties.set_prop("description", AAZStrType, ".description")
-                properties.set_prop("kind", AAZStrType, ".kind", typ_kwargs={"flags": {"required": True}})
+                properties.set_prop("kind", AAZStrType, ".type", typ_kwargs={"flags": {"required": True}})
                 properties.set_prop("onboarding", AAZObjectType, ".onboarding")
                 properties.set_prop("server", AAZObjectType, ".server")
                 properties.set_prop("title", AAZStrType, ".title", typ_kwargs={"flags": {"required": True}})
