+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/environment/_delete.py
@@ -19,7 +19,7 @@ class Delete(AAZCommand):
     """Delete the environment.
 
     :example: Delete environment
-        az apic environment delete -g api-center-test -s contosoeuap --name public
+        az apic environment delete -g api-center-test -s contosoeuap --environment-id public
     """
 
     _aaz_info = {
@@ -45,9 +45,9 @@ class Delete(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.environment_name = AAZStrArg(
-            options=["--name", "--environment", "--environment-name"],
-            help="The name of the environment.",
+        _args_schema.environment_id = AAZStrArg(
+            options=["--environment-id"],
+            help="The id of the environment.",
             required=True,
             id_part="child_name_2",
             fmt=AAZStrArgFormat(
@@ -126,7 +126,7 @@ class Delete(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "environmentName", self.ctx.args.environment_name,
+                    "environmentName", self.ctx.args.environment_id,
                     required=True,
                 ),
                 **self.serialize_url_param(
