+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/metadata/__cmd_group.py
@@ -8,11 +8,16 @@
 # pylint: skip-file
 # flake8: noqa
 
-from .__cmd_group import *
-from ._create import *
-from ._delete import *
-from ._export_metadata_schema import *
-from ._head import *
-from ._list import *
-from ._show import *
-from ._update import *
+from azure.cli.core.aaz import *
+
+
+@register_command_group(
+    "apic metadata",
+)
+class __CMDGroup(AAZCommandGroup):
+    """Commands to manage metadata schemas in API Center
+    """
+    pass
+
+
+__all__ = ["__CMDGroup"]
