+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/metadata/__init__.py
@@ -8,16 +8,11 @@
 # pylint: skip-file
 # flake8: noqa
 
-from azure.cli.core.aaz import *
-
-
-@register_command_group(
-    "apic metadata-schema",
-)
-class __CMDGroup(AAZCommandGroup):
-    """Commands to manage Metadata Schemas in API Center
-    """
-    pass
-
-
-__all__ = ["__CMDGroup"]
+from .__cmd_group import *
+from ._check_exists import *
+from ._create import *
+from ._delete import *
+from ._export import *
+from ._list import *
+from ._show import *
+from ._update import *
