+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/metadata/_check_exists.py
@@ -12,16 +12,16 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "apic api version head",
+    "apic metadata check-exists",
 )
-class Head(AAZCommand):
-    """Checks if specified API version exists.
+class CheckExists(AAZCommand):
+    """Checks if specified metadata schema exists.
     """
 
     _aaz_info = {
         "version": "2024-03-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.apicenter/services/{}/workspaces/{}/apis/{}/versions/{}", "2024-03-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.apicenter/services/{}/metadataschemas/{}", "2024-03-01"],
         ]
     }
 
@@ -41,11 +41,11 @@ class Head(AAZCommand):
         # define Arg Group ""
 
         _args_schema = cls._args_schema
-        _args_schema.api_name = AAZStrArg(
-            options=["--api", "--api-name"],
-            help="The name of the API.",
+        _args_schema.metadata_schema_name = AAZStrArg(
+            options=["--name", "--metadata-schema", "--metadata-schema-name"],
+            help="The name of the metadata schema.",
             required=True,
-            id_part="child_name_2",
+            id_part="child_name_1",
             fmt=AAZStrArgFormat(
                 max_length=90,
                 min_length=1,
@@ -64,32 +64,11 @@ class Head(AAZCommand):
                 min_length=1,
             ),
         )
-        _args_schema.version_name = AAZStrArg(
-            options=["--version", "--version-name"],
-            help="The name of the API version.",
-            required=True,
-            id_part="child_name_3",
-            fmt=AAZStrArgFormat(
-                max_length=90,
-                min_length=1,
-            ),
-        )
-        _args_schema.workspace_name = AAZStrArg(
-            options=["-w", "--workspace", "--workspace-name"],
-            help="The name of the workspace.",
-            required=True,
-            id_part="child_name_1",
-            default="default",
-            fmt=AAZStrArgFormat(
-                max_length=90,
-                min_length=1,
-            ),
-        )
         return cls._args_schema
 
     def _execute_operations(self):
         self.pre_operations()
-        self.ApiVersionsHead(ctx=self.ctx)()
+        self.MetadataSchemasHead(ctx=self.ctx)()
         self.post_operations()
 
     @register_callback
@@ -100,7 +79,7 @@ class Head(AAZCommand):
     def post_operations(self):
         pass
 
-    class ApiVersionsHead(AAZHttpOperation):
+    class MetadataSchemasHead(AAZHttpOperation):
         CLIENT_TYPE = "MgmtClient"
 
         def __call__(self, *args, **kwargs):
@@ -114,7 +93,7 @@ class Head(AAZCommand):
         @property
         def url(self):
             return self.client.format_url(
-                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ApiCenter/services/{serviceName}/workspaces/{workspaceName}/apis/{apiName}/versions/{versionName}",
+                "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ApiCenter/services/{serviceName}/metadataSchemas/{metadataSchemaName}",
                 **self.url_parameters
             )
 
@@ -130,7 +109,7 @@ class Head(AAZCommand):
         def url_parameters(self):
             parameters = {
                 **self.serialize_url_param(
-                    "apiName", self.ctx.args.api_name,
+                    "metadataSchemaName", self.ctx.args.metadata_schema_name,
                     required=True,
                 ),
                 **self.serialize_url_param(
@@ -143,14 +122,6 @@ class Head(AAZCommand):
                 ),
                 **self.serialize_url_param(
                     "subscriptionId", self.ctx.subscription_id,
-                    required=True,
-                ),
-                **self.serialize_url_param(
-                    "versionName", self.ctx.args.version_name,
-                    required=True,
-                ),
-                **self.serialize_url_param(
-                    "workspaceName", self.ctx.args.workspace_name,
                     required=True,
                 ),
             }
@@ -170,8 +141,8 @@ class Head(AAZCommand):
             pass
 
 
-class _HeadHelper:
-    """Helper class for Head"""
+class _CheckExistsHelper:
+    """Helper class for CheckExists"""
 
 
-__all__ = ["Head"]
+__all__ = ["CheckExists"]
